// src/1ak-auto/types.ts
export interface Product {
  brand?: string;
  name?: string;
  itemNumber?: string;
  price?: string;
  quantity?: string;
  delivery?: string;
}

export interface ProductToPreview {
  Бренд: string;
  Наименование: string;
  Артикул: string;
  Цена: string;
  Остаток: string;
  Поставка: string;
}